fix: File discovery for DunderVersion
There were multiple bugs with DunderVersion file discovery. First if it found a
version.py file but that file was not a "version" holding file, such as on
release-mgr itself, it would detect it and update nothing. Next, it would find
the test modules __init__.py and then also update nothing.

This new logic strips out obvious test modules and then makes sure the found
file has "__version__" in it before considering it found.